Make the rank two update more efficient.
Refactor the rank two update into a separaete file of its own.
Also modify the timing code so that we don't count the time
taken for the preconditioning. This version can't be used for final
reporting. This is only to get an idea of where exactly things are
going wrong. It seems that the preconditioning is too expensive but
otherwise even at 50 dimensions I am able to bypass the vanilla 2SPSA.